Added implementation for updateProxies and added updateParserProxiesJob

+++ src/main/resources/db/migration/V1.0.0__init.sql
@@ -8,5 +8,5 @@ CREATE TABLE parser_proxies
 (
     id bigint PRIMARY KEY,
     parser_name varchar(50),
-    ping float
+    ping bigint
 )
